azure_functions/train/requirements.txt
azure-functions==1.13.3
pandas
scikit-learn
